backend/requirements.txt
fastapi[standard]
redis
requests
python-dotenv
geocoder
google-genai
sentence-transformers
